[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -3,11 +3,11 @@
 using namespace std;
 
 int main() {
-	int a, b;
+	long long a, b;
 	cin >> a >> b;
 
 	if (a > b) {
-		int tmp = a;
+		long long tmp = a;
 		a = b;
 		b = tmp;
 	}
@@ -15,7 +15,7 @@ int main() {
 	if (a == b) cout << "0\n";
 	else cout << b - a - 1 << "\n";
 
-	for (int i = a + 1;i < b;i++) cout << i << " ";
+	for (long long i = a + 1;i < b;i++) cout << i << " ";
 
 	return 0;
 }
